backend/requirements.txt
Flask==1.1.2
Werkzeug==1.0.1
jinja2==2.11.3
MarkupSafe==2.0.1
psycopg2-binary==2.9.3
itsdangerous==2.0.1
